+++ src/tracks/Argument.java
@@ -6,10 +6,12 @@ package tracks;
  * arg: colour to use
  * invert: invert matching? */
 class Argument {
+	private String key;
 	private String arg;
 	private boolean invert= false;
 	
-	Argument(String arg, boolean invert){
+	Argument(String key, String arg, boolean invert){
+		this.setKey(key);
 		this.setArg(arg);
 		this.setInvert(invert);
 	}
@@ -20,10 +22,18 @@ class Argument {
 	protected void setArg(String arg) {
 		this.arg = arg;
 	}
+	
 	protected boolean isInvert() {
 		return invert;
 	}
 	protected void setInvert(boolean invert) {
 		this.invert = invert;
+	}
+
+	protected String getKey() {
+		return key;
+	}
+	protected void setKey(String key) {
+		this.key = key;
 	}
 }