+++ test/tracks/GenotypeMatrixTest.java
@@ -6,11 +6,16 @@ import java.io.File;
 import java.io.IOException;
 import java.sql.SQLException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
+
+import javax.script.ScriptEngine;
+import javax.script.ScriptEngineManager;
+import javax.script.ScriptException;
 
 import org.junit.Before;
 import org.junit.Test;
+
+import com.google.common.base.Joiner;
 
 import coloring.Config;
 import coloring.Xterm256;
@@ -18,7 +23,6 @@ import exceptions.InvalidColourException;
 import exceptions.InvalidConfigException;
 import exceptions.InvalidGenomicCoordsException;
 import exceptions.InvalidRecordException;
-import htsjdk.variant.variantcontext.VariantContext;
 import htsjdk.variant.vcf.VCFFileReader;
 import htsjdk.variant.vcf.VCFHeader;
 import samTextViewer.GenomicCoords;
@@ -31,20 +35,45 @@ public class GenotypeMatrixTest {
 		new Xterm256();
 	}
 
-//	@Test
-//	public void bigData()throws InvalidGenomicCoordsException, IOException, ClassNotFoundException, InvalidRecordException, SQLException, InvalidColourException {
-//
-//		GenomicCoords gc= new GenomicCoords("1:572807-755079", 200, null, null);
-//		TrackIntervalFeature vcf= new TrackIntervalFeature("/Users/db291g/Downloads/ALL.wgs.mergedSV.v8.20130502.svs.genotypes.vcf.gz", gc);
-//		vcf.setNoFormat(false);
-//		long t0= System.currentTimeMillis();
-//		System.err.println(vcf.printToScreen().length());
-//		long t1= System.currentTimeMillis();
-//		System.err.println(t1-t0);
-//	}	
+	@Test
+	public void testEval() throws ScriptException, InvalidColourException, InvalidGenomicCoordsException, IOException, ClassNotFoundException, InvalidRecordException, SQLException{
+		String f= "/Users/db291g/Downloads/ALL.wgs.mergedSV.v8.20130502.svs.genotypes.vcf.gz";
+
+		VCFFileReader reader = new VCFFileReader(new File(f));
+		VCFHeader vcfHeader = reader.getFileHeader();
+		reader.close();
+		GenomicCoords gc= new GenomicCoords("1:5-35444567", 80, null, null);
+		TrackIntervalFeature vcf= new TrackIntervalFeature(f, gc);
+
+		List<IntervalFeature> linf = vcf.getIntervalFeatureList();
+
+		GenotypeMatrix gm= new GenotypeMatrix();
+		gm.setJsScriptFilter("{AC} < 0");
+		long t0= System.currentTimeMillis();
+		gm.printToScreen(true, linf, 80, vcfHeader);
+		long t1= System.currentTimeMillis();
+		System.err.println(t1-t0);
+		
+	}
+	
+	@Test
+	public void bigData()throws InvalidGenomicCoordsException, IOException, ClassNotFoundException, InvalidRecordException, SQLException, InvalidColourException {
+
+		VCFFileReader reader = new VCFFileReader(new File("test_data/ALL.wgs.mergedSV.v8.20130502.svs.genotypes.vcf.gz"));
+		VCFHeader vcfHeader = reader.getFileHeader();
+		reader.close();
+		GenomicCoords gc= new GenomicCoords("1:5934301-6000000", 80, null, null);
+		TrackIntervalFeature vcf= new TrackIntervalFeature("test_data/ALL.wgs.mergedSV.v8.20130502.svs.genotypes.vcf.gz", gc);
+
+		List<IntervalFeature> linf = vcf.getIntervalFeatureList();
+
+		GenotypeMatrix gm= new GenotypeMatrix();
+		gm.setJsScriptFilter("{GT} == 10 || 1>2");
+		gm.printToScreen(true, linf, 80, vcfHeader);
+	}	
 
 	@Test
-	public void canFilterWithJavascript() throws InvalidGenomicCoordsException, IOException, ClassNotFoundException, InvalidRecordException, SQLException, InvalidColourException {
+	public void canFilterWithJavascript() throws InvalidGenomicCoordsException, IOException, ClassNotFoundException, InvalidRecordException, SQLException, InvalidColourException, IOException {
 		
 		VCFFileReader reader = new VCFFileReader(new File("test_data/info_formats.vcf.gz"));
 		VCFHeader vcfHeader = reader.getFileHeader();
@@ -60,39 +89,188 @@ public class GenotypeMatrixTest {
 		// No filter
 		assertEquals(2, gm.printToScreen(true, linf, 80, vcfHeader).split("\n").length);
 
-		// 
+		// Filter by array in FORMAT
+		gm.setJsScriptFilter("{FF}[1] > 0");
+		String y= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertTrue(y.contains("sample1"));
+		assertTrue( ! y.contains("sample2") );
+		
+		// One sample satisfies DP at at least one record.
 		gm.setJsScriptFilter("{DP} >= 100");
 		String x= gm.printToScreen(true, linf, 80, vcfHeader);
 		assertEquals(1, x.split("\n").length);
+
+		// Filter by array in INFO
+		gm.setJsScriptFilter("{XA}[0] > 0");
+		x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertTrue(x.contains("sample1"));
+
+		// Exclude all samples.
+		gm.setJsScriptFilter("{DP} >= 10000");
+		x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertTrue(x.isEmpty());
+
+		// Exclude all sample II: Non sense comparison
+		gm.setJsScriptFilter("{POS} == 'G' && {DP} > 5");
+		x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertTrue(x.isEmpty());
+
+		// Remove filter
+		gm.setJsScriptFilter("1 > 2"); // First remove all samples
+		x= gm.printToScreen(true, linf, 80, vcfHeader); 
+		assertTrue(x.isEmpty());
 		
+		gm.setJsScriptFilter(null); // Then remove filter and check samples are back
+		x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertTrue( ! x.isEmpty());
+		
+		gm.setJsScriptFilter(""); // Empty also works to remove filter
+		x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertTrue( ! x.isEmpty());
+		
+		// Handle missing values
+		gm.setJsScriptFilter("{ID} == '.'"); // Use literal '.', not the null value.
+		x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertTrue( x.trim().length() > 50);
+
+		// Filter by FLAG field type
+		gm.setJsScriptFilter("{XB} && {DP} == 100");
+		x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertTrue(x.contains("sample2"));
+		assertTrue( ! x.contains("sample1"));
+
 		// Filter by ALT: Note array slicing
 		// NB: Although all samples satisfy ALT, only sample2 also satisfies DP
 		gm.setJsScriptFilter("{ALT}[0] == 'G' && {DP} > 5");
 		x= gm.printToScreen(true, linf, 80, vcfHeader);
 		assertEquals(1, x.split("\n").length);
+		assertTrue(x.contains("sample2"));
 
-		// JS script must return boolean
-		boolean pass= false;
-		try{
-			gm.setJsScriptFilter("10 + 3");
-			gm.printToScreen(true, linf, 80, vcfHeader);
-		} catch(InvalidGenomicCoordsException e){
-			pass= true;
-		}
-		assertTrue(pass);
+		// Same using POS
+		gm.setJsScriptFilter("{POS} == 17822094 && {DP} > 5");
+		x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertEquals(1, x.split("\n").length);
+		assertTrue(x.contains("sample2"));
+	}
 
-		GenotypeMatrix gm2= new GenotypeMatrix();
-		int i= 0;
-		while(i < 1000){
-			gm2.setJsScriptFilter("{ALT}[0] == 'G' && {DP} > 5");
-			String y= gm2.printToScreen(true, linf, 80, vcfHeader);
-			System.err.println(i);
-			i++;
-		}
+	@Test
+	public void genotypeAsNumericInJS() throws ClassNotFoundException, IOException, InvalidGenomicCoordsException, InvalidRecordException, SQLException, InvalidColourException, IOException{
+		VCFFileReader reader = new VCFFileReader(new File("test_data/info_formats.vcf.gz"));
+		VCFHeader vcfHeader = reader.getFileHeader();
+		reader.close();
+		GenomicCoords gc= new GenomicCoords("1:17822074-17822184", 80, null, null);
+		TrackIntervalFeature vcf= new TrackIntervalFeature("test_data/info_formats.vcf.gz", gc);
+		List<IntervalFeature> linf = vcf.getIntervalFeatureList();
+
+		GenotypeMatrix gm= new GenotypeMatrix();
+		
+		gm.setJsScriptFilter("{GT} == '1/0'");
+		String x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertTrue(x.contains("sample1"));
+		assertTrue( ! x.contains("sample2") );
+		
+		gm.setJsScriptFilter("{GT} == './.'");
+		x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertTrue( x.contains("sample1"));
+		assertTrue( ! x.contains("sample2") );
 	}
 	
 	@Test
-	public void canInitMatrix() throws InvalidGenomicCoordsException, IOException, ClassNotFoundException, InvalidRecordException, SQLException, InvalidColourException {
+	public void genotypeKetwordInJS() throws ClassNotFoundException, IOException, InvalidGenomicCoordsException, InvalidRecordException, SQLException, InvalidColourException, IOException{
+
+		VCFFileReader reader = new VCFFileReader(new File("test_data/info_formats.vcf.gz"));
+		VCFHeader vcfHeader = reader.getFileHeader();
+		reader.close();
+		GenomicCoords gc= new GenomicCoords("1:17822074-17822184", 80, null, null);
+		TrackIntervalFeature vcf= new TrackIntervalFeature("test_data/info_formats.vcf.gz", gc);
+		List<IntervalFeature> linf = vcf.getIntervalFeatureList();
+
+		GenotypeMatrix gm= new GenotypeMatrix();
+		
+		gm.setJsScriptFilter("{HOM}");
+		String x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertTrue(x.contains("sample2"));
+		assertTrue( ! x.contains("sample1") );
+		
+		gm.setJsScriptFilter("{NO_CALL}");
+		x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertTrue(x.contains("sample1"));
+		assertTrue( ! x.contains("sample2") );
+		
+		gm.setJsScriptFilter("{HET_NON_REF}");
+		x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertTrue(x.contains("sample2"));
+		assertTrue( ! x.contains("sample1") );
+		
+		gm.setJsScriptFilter("{CALLED} && {POS} == 17822094");
+		x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertTrue(x.contains("sample2"));
+		assertTrue( ! x.contains("sample1") );
+		
+	}
+	
+	@Test
+	public void invalidJSscripts() throws ClassNotFoundException, IOException, InvalidGenomicCoordsException, InvalidRecordException, SQLException, InvalidColourException, IOException{
+
+		VCFFileReader reader = new VCFFileReader(new File("test_data/info_formats.vcf.gz"));
+		VCFHeader vcfHeader = reader.getFileHeader();
+		reader.close();
+		GenomicCoords gc= new GenomicCoords("1:17822074-17822184", 80, null, null);
+		TrackIntervalFeature vcf= new TrackIntervalFeature("test_data/info_formats.vcf.gz", gc);
+		List<IntervalFeature> linf = vcf.getIntervalFeatureList();
+
+		GenotypeMatrix gm= new GenotypeMatrix();
+		
+		// JS script must return boolean
+		gm.setJsScriptFilter("{DP} > 5 && 10 + 3");
+		gm.printToScreen(true, linf, 80, vcfHeader);
+		assertNull(gm.getJsScriptFilter()); // After faulty script reset to null. 
+
+		// Invalid JS syntax. E.g. when {TAG} does not exist.
+		gm.setJsScriptFilter("{FOOBAR} > 10");
+		gm.printToScreen(true, linf, 80, vcfHeader);
+		assertNull(gm.getJsScriptFilter());
+		
+		// After exception the invalid filter has been removed
+		gm.printToScreen(true, linf, 80, vcfHeader);
+	}
+	
+	@Test
+	public void canFilterGenotypeWithJS() throws IOException, ClassNotFoundException, IOException, InvalidGenomicCoordsException, InvalidRecordException, SQLException, InvalidColourException{
+
+		VCFFileReader reader = new VCFFileReader(new File("test_data/info_formats.vcf.gz"));
+		VCFHeader vcfHeader = reader.getFileHeader();
+		reader.close();
+		GenomicCoords gc= new GenomicCoords("1:17822074-17822184", 80, null, null);
+		TrackIntervalFeature vcf= new TrackIntervalFeature("test_data/info_formats.vcf.gz", gc);
+		List<IntervalFeature> linf = vcf.getIntervalFeatureList();
+
+		GenotypeMatrix gm= new GenotypeMatrix();
+		
+//		// NB: The JS must have the allele bases REF and ALT, not the encoding 
+//		// 0/1, 1/1 etc. that you see in the VCF
+//		
+//		// Note '/' separator (same as in VCF)
+//		gm.setJsScriptFilter("{GT} == 'G/T'");
+//		String x= gm.printToScreen(true, linf, 80, vcfHeader);
+//		assertTrue(x.contains("sample2"));
+//		assertTrue( ! x.contains("sample1"));
+//		
+//		// Here '|' separator (same as in VCF)
+//		gm.setJsScriptFilter("{GT} == 'T|A'");
+//		x= gm.printToScreen(true, linf, 80, vcfHeader);
+//		assertTrue(x.contains("sample2"));
+//		assertTrue( ! x.contains("sample1"));
+
+		// Missing alleles
+		gm.setJsScriptFilter("{GT} == './.'");
+		String x= gm.printToScreen(true, linf, 80, vcfHeader);
+		assertTrue(x.contains("sample1"));
+		assertTrue( ! x.contains("sample2"));
+	}
+	
+	@Test
+	public void canInitMatrix() throws IOException, InvalidGenomicCoordsException, IOException, ClassNotFoundException, InvalidRecordException, SQLException, InvalidColourException {
 		
 		VCFFileReader reader = new VCFFileReader(new File("test_data/ALL.wgs.mergedSV.v8.20130502.svs.genotypes.vcf.gz"));
 		VCFHeader vcfHeader = reader.getFileHeader();
@@ -157,7 +335,7 @@ public class GenotypeMatrixTest {
 	}
 	
 	@Test
-	public void canHandleIntervalWithNoFeatures() throws InvalidGenomicCoordsException, IOException, ClassNotFoundException, InvalidRecordException, SQLException, InvalidColourException{
+	public void canHandleIntervalWithNoFeatures() throws IOException, InvalidGenomicCoordsException, IOException, ClassNotFoundException, InvalidRecordException, SQLException, InvalidColourException{
 		GenomicCoords gc= new GenomicCoords("2:1-1000", 80, null, null);
 		TrackIntervalFeature vcf= new TrackIntervalFeature("test_data/ALL.wgs.mergedSV.v8.20130502.svs.genotypes.vcf.gz", gc);
 		List<IntervalFeature> linf = vcf.getIntervalFeatureList();