+++ test/tracks/TextReadTest.java
@@ -7,6 +7,7 @@ import static org.junit.Assert.assertTrue;
 import java.io.File;
 import java.io.IOException;
 
+import org.junit.Before;
 import org.junit.Test;
 
 import coloring.Config;
@@ -27,6 +28,11 @@ public class TextReadTest {
 	public static SAMSequenceDictionary samSeqDict= samReader.getFileHeader().getSequenceDictionary();
 	
 	public static String fastaFile= "test_data/chr7.fa";
+	
+	@Before
+	public void setConfig() throws IOException, InvalidConfigException{
+		new Config(null);
+	}
 	
 	@Test
 	public void getNoAlignedBasesFromCigar() {
@@ -71,9 +77,7 @@ public class TextReadTest {
 	}
 
 	@Test
-	public void canPrintDNARead() throws InvalidGenomicCoordsException, IOException, InvalidColourException, InvalidConfigException{
-		
-		new Config(null);
+	public void canPrintDNARead() throws InvalidGenomicCoordsException, IOException, InvalidColourException {
 		
 		GenomicCoords gc= new GenomicCoords("chr7:5566778-5566798", 80, samSeqDict, fastaFile);
 		SAMRecord rec= new SAMRecord(null);